Improved performance by commenting out unnecessary scripts
The asynchronous script 'https://events.framer.com/script' was commented out in the 'smspulsera.framer.ai_443/index.html' file as it was found to be causing a performance issue on the website. The script was not necessary for the functioning of the website. By commenting it out, it ensures that it does not load and hence, increases the speed and performance of the site.